refactor(navbar):seprated navbar into smaller cp's
