+++ src/Styles/Appbar/index.tsx
@@ -1,0 +1,38 @@
+import { styled } from "@mui/material/styles";
+import { Tab, Tabs } from "@mui/material";
+
+interface StyledTabProps {
+  label: string;
+}
+
+export const AntTab = styled((props: StyledTabProps) => (
+  <Tab disableRipple {...props} />
+))(({ theme }) => ({
+  textTransform: "capitalize",
+  minWidth: 0,
+  [theme.breakpoints.up("sm")]: {
+    minWidth: 0,
+  },
+  fontWeight: theme.typography.fontWeightMedium,
+  fontSize: 15,
+  marginRight: theme.spacing(1),
+  color: "rgba(0, 0, 0, 0.85)",
+  borderBottom: "3px solid white",
+
+  "&:hover": {
+    color: "#f03637",
+    borderBottom: "3px solid #f03637",
+  },
+  "&.Mui-selected": {
+    color: "#333333",
+  },
+}));
+
+export const AntTabs = styled(Tabs)(({ theme }) => ({
+  "& .MuiTabs-indicator": {
+    backgroundColor: "#f03637",
+  },
+  [theme.breakpoints.down("md")]: {
+    display: "none",
+  },
+}));
